Update predict.md
There is an issue with the [predict docs](https://docs.ultralytics.com/predict/) for the Python SDK. `results` is a list of `torch.Tensor` objects. Running a `for` loop and then saying `boxes = results.boxes` is incorrect and returns an error accordingly. Because the attribute `boxes` is valid for each `result` object and not `List[result]`.