Remove hard-coded batch size of 1 from Exporter
This line prevents users from providing a custom batch size to the exporter with kwargs in `model.export`. When it's removed, providing a batch size works as expected. I verified this by exporting one of my v8n models to onnx and providing a custom batch size, then opening the model on netron.app.
```
from ultralytics import YOLO

model = YOLO("trained_models/v8n.pt")

model.export(format="onnx", batch=8)
```

It looks like this is an artifact from a work-around introduced in #129. In #132 the conditional was commented out but the body was left in.