Add preload function for fast execution when `predict` is called
#### Current issue
When calling `predict`, it takes a few seconds to build the predictor & run `setup_model`. When running in a pipeline, it's ideal to have the everything loaded and ready to go when a request comes in.

#### Proposed solution
Add a `preload` function which simply creates the predictor without having to call `predict` first and wait.