Update version of PyTorch <2.0
This pull request is to address the bug in the latest version of PyTorch, which produces the following error message:

"File "/opt/conda/envs/.../lib/python3.10/site-packages/torch/nn/modules/conv.py", line 459, in _conv_forward
return F.conv2d(input, weight, bias, self.stride,
RuntimeError: GET was unable to find an engine to execute this computation"

After some investigation, I found that this error is caused by a compatibility issue between PyTorch and some other libraries used in the project. Specifically, the PyTorch version used in the project is not compatible with the other dependencies, which leads to the above error.

To solve this issue, I propose enforcing the PyTorch package to be in the version <2.0, and Torchvision <0.15, which is compatible with the other dependencies used in the project. I have tested this solution on my local machine and verified that it resolves the error.

Therefore, I suggest updating the project requirements to include:

torch>=1.7.0,<2.0
torchvision>=0.8.1,<0.15

I hope this solution will be helpful in resolving the issue, and I am happy to provide any further assistance.