clearml: show training progress after each epoch
Currently we only display the training progress only after the training
has finished. This has an important downside for long training jobs,
when failure to react in a suboptimal training, can lead to huge compute
resource wast and significant costs.

This commits adds a scalar plot, which is updated at the end of every
epoch. It includes the following metrics:
- fitness
- mAP50
- mAP50-95
- precision
- recall

Signed-off-by: Evangelos Petrongonas <e.petrongonas@hellenicdrones.com>