Update exporter.py
Changed coreml export extension to a working format. From ".mlmodel" to ".mlpackage".

reason:
If we leave it as .mlmodel, the following error message will appear because the pipeline model with nms added cannot be converted. "Exception: For an ML Program, extension must be .mlpackage (not .mlmodel)"

The .mlpackage format works fine.

Updated export.py to save Core ML models in the new .mlpackage format for better compatibility and flexibility.

mlpackage format
more flexible and compatible
cutting edge for Core ML

Walkthrough
Change the file extension for saving the Core ML model to .mlpackage to support the new format in Core ML Tools 5.0