Fix desync between update thread and consumer.
Found the cause of the weird errors on the timings.
When iterating over ´LoadStreams´ a copy is handled with whatever the update thread has at the moment.
The problem for example was that the main thread was slower in asking for copys of the frames produced by the update thread, so frames in-between where lost, and worse, if the main thread was faster (for example a network delay on the update), it could copy the same frame it was already on the consumed variable.
Importing the thread library and using conditions solves this with no delay.

While profiling the consumer ´BasePredictor´ tough, I'm seeing a 20% extra delay on things that are not pre,post, or inferring phases or callbacks, somewhere i have not found yet is causing extra delays on the whole process.