Notify all threads if there are multiple sources.
The whole imgs array is copied on __next__ so that means all of the update threads must be in sync.
The problem with this design is that if one update thread has delay while others not, when the main threads copies all the produced values from all update threads, the one with problems will be duplicated.