Fix extra delay for double waitkey call on streams.
Found the extra delay (16ms/20%) on my system, a lot, caused by a waitKey on "q" to quit check made inside the iterator, moved it to where imgshow is.
The problem is that this waitKey takes a lot of time off the main thread even when no show command is false.
Moving it to where show is fixes the extra delay because a waitkey was already made there, also fixes the lag when pressing q to quit when show=true.
Also, the q key now works on all show MODES, previously it only worked on streams with some lag.