Fix live streams.
Totally forgot that LoadStream also is for cameras and live streams when refactored remote files a few commits ago.
Now i see the original intent of this class and why the daemons where not in sync.

* Restored the original code to handle live streams and added support for non live streams.

*Also fix youtube live stream when using yt-dlp

We could do a refactor for non live streams or leave it like this, need guidance.