Fixes Result.tojson when tracking
The code works for predictions but the tracking data row contains a new value on the fifth position (row[4]) which is the id. This results in output such as:

```
  {
    "name": "person",
    "class": 0,
    "confidence": 295.0,
    "box": {
      "x1": 1003.8906860351562,
      "y1": 93.98287200927734,
      "x2": 1060.8193359375,
      "y2": 249.63829040527344
    }
  }
```

The confidence returned is actually the tracking id and the class only returns 0 because it's casting the confidence (0...0.999) to int, which always returns 0.