Make OpenVINO device be configurable
OpenVINO AUTO device usually starts running on CPU and after the model is ready on another device (iGPU in my case), switches to it. It results in unstable predictions because implementation varies from device to device.
The problem with the proposed solution is that `device` comes from `select_device()` which requires it to be one of the torch devices. That prohibits having AUTO.

Close https://github.com/ultralytics/ultralytics/issues/4721