🐞 fix(ultralytics/utils/callbacks/base.py add_integration_callbacks): callbacks coverage
In the add_integration_callbacks, using the update method of a dictionary can cause callbacks to be overwritten. Such as the callbacks of tensorboard will overwrite the callbacks of comet.
And the callbacks variable should be a dictionary with a list as its value.
To fix this bug, use a callback_ list to temporarily save all callbacks, and update the instance's callbacks dictionary at the end.