Update tuner.py to enable hyperparameter tuning via Ray Tune for larger models
Hello! :) 

Here I utilized ray.put() so that model passed inside a trainable is inside ray store. That way, the problem, as indicated by the error "ValueError: The actor ImplicitFunc is too large", and which occurs for models other than "nano", is solved.

In by putting the larger models in ray store and "unpacking" via ray.get() inside the trainable function, the hyperparameter search can take place for larger models.