Account for machine epsilon in image _single_check
The previous comparison, im.max() > 1.0, was failing on my image with the message:

```
WARNING  torch.Tensor inputs should be normalized 0.0-1.0 but max value is 1.0000001192092896. Dividing input by 255.
```

However, my image was fine; this number is actually 1, just within the machine epsilon of the torch.float32 data type.

This new check accounts for that properly.