Repair converter.py box-segment discarding
The previous code checked and discarded identical bboxes, but then continued processing the label and made the same check with the segmentation. This means that for some cases where the boxes are slightly different, but the segmentation coincides (due to auto-labelling or other reasons), the system discards the segment and ends up with N bboxes and N-1 segments. This generates an out-of-bounds error in line 144. This check is not performed for keypoints, thus the code has been modified to make the discarding  of the whole label (bbox, segment, and keypoint) only with the bbox check.