Update docstrings of mouse_callback function for Homework purpose
Dear Reviewer,

At the beginning of the semester for our Software Architecture lecture, we were tasked with selecting a project to analyze. After careful consideration, we chose a project that met the specified requirements. As we approach the end of the semester, our final assignment involves making a valid change to the selected project, committing and pushing the modification, and communicating with the project owner to discuss our updates.

Given that our team does not have specific requirements related to mathematical or other technical changes, we decided to modify a function's docstring within the project. This adjustment serves as our contribution to the project.

This request is part of our homework, and we appreciate your time in reviewing and providing feedback on our changes.

Thank you for your consideration.

Sincerely,
Orhan Murat Tuncer, Cansu Zeynep Kars, Ece Büşra Civelek, Fatma Sena Alpakut