Fix for hsv augmentation
Currently, colors, which in HSV notation, have a small hue value, are practically not affected by hue value change augmentation. This is because the new hue values are determined by multiplying the current ones by a certain scalar. In the commited version, they are shifted the same way regardless of the initial value.