Hybrid Agnostic NMS
Compute agnositc NMS considering a subset of classes, and non-agnostic
NMS for the rest of the classes.

This is useful when some classes should not be colocated, while others
should be.

For example, take a model trained to detect people and several classes
of shirt, where each shirt class is trained on a different shirt color.
In this case, we'd like to detect only a single shirt class per person
by using agnostic NMS on just the shirt classes, while allowing the
person class to be detected with substantial overlap with any shirt
class.

This is accomplished by splitting the predictions during postprocessing,
computing agnostic NMS for the shirts, and non-agnostic NMS for the
person. This approach is implemented in this PR.