Using a zero tensor to run warmup instead of empty tensor
Though there is nothing wrong with pushing an empty tensor through the
network, there are failure cases or cases with unwanted side-effects
associated with it. The torch.empty() call does not necessarily
initialize the tensor.
Also see https://pytorch.org/docs/stable/generated/torch.empty.html

An unwanted side-effect can occur if the values that are found in the
tensor are wildly out of range of the usual values for an image or
are NaN in the worst-case scenario. Normally, the output of the
warmup runs are irrelevant and pushing NaNs is fine, however in
cases were _stateful_ layers like BatchNorm are used, this will
destroy the running statistics and cause unexpected results in
subsequent runs where the results are important. Pushing a
zero tensor here has negligible impact on the running statistics
of BatchNorm layers.

When disabling fusing of the BatchNorm to the Conv layers, I observed
signficant drops in the metrics after the final training epoch due to
the effect described above.

Note that there are a few other instances in the codebase that
use empty tensors (autobatch.py and torch_utils.py) but those runs
are used for either training batch determination prior to training
or to compute model FLOPs. For good mesaure I recommend setting
the tensor to a zero tensor on those files too.