Ensure matplotlib backend gets reset with plt_settings
The plt_settings decorator attempts to restore the matplolib settings
after the decorated function runs.  However, if the decorated function
raises an error, it will not entirely succeed.  The rc_context
manager will reset most settings, but it does not handle the backend.
We need to ensure the backend reset code will run, by sticking it in
a finally block.

Also, this makes a single check if the backend will be changing, since
the two checks that were here were slightly inconsistent.