fix OBB train bug
The cv2.minAreaRect() function returns the rotation angle θ, which is the counterclockwise rotation from the x-axis to the first encountered edge of the rectangle. The length of this edge is w, and the length of the other edge is h. Thus, the range of θ is (0,90]. Therefore, when training an Oriented Bounding Box (OBB) detection model, a rectangle that was originally horizontal will be transformed into a vertical rectangle. As a result, the model ultimately learns a vertical rectangle. In OBB training, it is necessary to ensure that the handling of angles is consistent with the model's expectations to prevent the model from learning the wrong direction of the box.