Updated to work on my machine
Changed the toxml() and tocsv() methods to use pd.dataframe instead of the todataframe() method. This works on my machine but this is in an idealised testing environment. I cant work out how to use this branch as an import so if I could get some guidance with that it would be very helpful