Update pyproject.toml to include lxml
When testing the pr from #16267 (thank you @ambitious-octopus for the advice) I ran into a problem with a completely fresh install that the `lxml` package isnt installed with the ultralytics package which then was installed automatically but needed me to rerun the script. It might be worth adding `lxml` to `pyproject.toml` to prevent this from needing to be done every time that someone new runs the method.